Added captcha validation for password recovery form.
Details:
http://www.myjqueryplugins.com/jquery-plugin/qaptcha
When captcha is completed, a post request is sent to the QapTchaServlet, which sets 'qaptcha_key' session attribute. The attribute is then used during input validation in managed bean. Captcha is considered to be validated if in request parameters there is an empty parameter with name = session['qaptcha_key'].

To add a captcha in any form, just add an empty div or span with class="QapTcha".